layout save and code alignment changes
