+++ app/lib/refs_config.json
@@ -1,1 +1,1 @@
-[{"_id":"refs", "ref_ids":[{"ref_id":"eng_ulb", "ref_name":"English-ULB", "isDefault":true}, {"ref_id":"eng_udb", "ref_name":"English-UDB", "isDefault":false}, {"ref_id":"hin_ulb", "ref_name":"Hindi-ULB", "isDefault":false}, {"ref_id":"arb_vdt", "ref_name":"Arabic-VDT", "isDefault":false}]}, {"_id":"ref_history", "visit_history":[{"book":"Genesis", "chapter":"1", "bookId": "1"}]}, {"_id": "autoupdate", "enable": false, "updateDownload": false}, {"_id": "app_locale", "appLang": "en"}, {"_id": "appFirstRun", "isFirstRun": true }, {"_id": "activeRefs", "activeRefs": {"0": "eng_ulb", "1": "eng_ulb", "2": "eng_ulb"}}]
+[{"_id":"refs", "ref_ids":[{"ref_id":"eng_ulb", "ref_name":"English-ULB", "isDefault":true}, {"ref_id":"eng_udb", "ref_name":"English-UDB", "isDefault":false}, {"ref_id":"hin_ulb", "ref_name":"Hindi-ULB", "isDefault":false}, {"ref_id":"arb_vdt", "ref_name":"Arabic-VDT", "isDefault":false}]}, {"_id":"ref_history", "visit_history":[{"book":"Genesis", "chapter":"1", "bookId": "1"}]}, {"_id": "autoupdate", "enable": false, "updateDownload": false}, {"_id": "app_locale", "appLang": "en"}, {"_id": "appFirstRun", "isFirstRun": true }, {"_id": "activeRefs", "activeRefs": {"0": "eng_ulb", "1": "eng_ulb", "2": "eng_ulb"}}, {"_id": "targetReferenceLayout", "layout": 1}]
